updated client folders
